This work was done over ~80 individual commits in the `cmake` branch.

BEGIN_COMMIT_OVERRIDE
feat!: Replace gyp build system with CMake

Closes #1047

feat!: Update all dependencies
fix: Fix Python 3.10+ compatibility in scripts
feat!: Drop Python 2 support in all scripts
feat!: Replace glog with absl::log, tweak log output and flags
feat: Add install target to build system
feat: Respect the file mode for HttpFiles (#1081)
feat: Move all third-party deps into git submodules (#1083)
fix: Fix local files with UTF8 names (#1246)
fix: Fix build on FreeBSD (#1287)
fix: Fix clang build (#1288)
fix: Fix various build issues on macOS
fix: Fix various build issues on Windows
END_COMMIT_OVERRIDE

+++ packager/media/base/video_stream_info.h
@@ -1,4 +1,4 @@
-// Copyright 2014 Google Inc. All rights reserved.
+// Copyright 2014 Google LLC. All rights reserved.
 //
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file or at
@@ -7,7 +7,7 @@
 #ifndef PACKAGER_MEDIA_BASE_VIDEO_STREAM_INFO_H_
 #define PACKAGER_MEDIA_BASE_VIDEO_STREAM_INFO_H_
 
-#include "packager/media/base/stream_info.h"
+#include <packager/media/base/stream_info.h>
 
 namespace shaka {
 namespace media {
@@ -35,8 +35,8 @@ class VideoStreamInfo : public StreamInfo {
                   const std::string& codec_string,
                   const uint8_t* codec_config,
                   size_t codec_config_size,
-                  uint16_t width,
-                  uint16_t height,
+                  uint32_t width,
+                  uint32_t height,
                   uint32_t pixel_width,
                   uint32_t pixel_height,
                   uint8_t transfer_characteristics,
@@ -56,8 +56,8 @@ class VideoStreamInfo : public StreamInfo {
 
   const std::vector<uint8_t>& extra_config() const { return extra_config_; }
   H26xStreamFormat h26x_stream_format() const { return h26x_stream_format_; }
-  uint16_t width() const { return width_; }
-  uint16_t height() const { return height_; }
+  uint32_t width() const { return width_; }
+  uint32_t height() const { return height_; }
   /// Returns the pixel width.
   /// @return 0 if unknown.
   uint32_t pixel_width() const { return pixel_width_; }
@@ -100,8 +100,8 @@ class VideoStreamInfo : public StreamInfo {
   // to mp4 container only. It is needed by some codecs, e.g. Dolby Vision.
   std::vector<uint8_t> extra_config_;
   H26xStreamFormat h26x_stream_format_;
-  uint16_t width_;
-  uint16_t height_;
+  uint32_t width_;
+  uint32_t height_;
 
   // pixel_width_:pixel_height_ is the sample aspect ratio.
   // 0 means unknown.
